+++ extensions/bunch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Bunch Changelog
 
+## [Add new command: Create Bunches] - 2022-06-18
+
+- Add new command: Create Bunches.
+
 ## [Initial Version] - 2022-06-15
 
--Control [Bunch](https://bunchapp.co) via Raycast.
+- Control [Bunch](https://bunchapp.co) via Raycast.